+++ pyyabc/element/base.py
