+++ pyyabc/molecular/all_perms.txt
