+++ pyyabc/utils/math.py
