+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='pyabc',
+    name='pyyabc',
     # Remenber change version in ababe.cmdline.runabalib.exec_from_cmdline()
     version='0.1.0',
     description='A grid structure machine learning tool used\
@@ -23,7 +23,7 @@ setup(
     scripts=[],
     entry_points={
         'console_scripts': [
-            'rexpand = pyabc.cmd.rexpand:cli',
+            'rexpand = pyyabc.cmd.rexpand:cli',
         ],
     }
 )
